Quickfixes for float32 -> float64 conversions

+++ src/predict.jl
@@ -51,7 +51,7 @@ function predict(
             # Calculate apparent sky coordinates for this timestep
             # Measurement set columns are in modified Julian date, but in *seconds*.
             # We hardcode the location of MWA for now.
-            frame = Frame(time / (24 * 3600), 2.0362898433426406, -0.46606084551737853)
+            frame = Frame(Float64(time / (24 * 3600)), 2.0362898433426406, -0.46606084551737853)
             for comp in comps
                 alt, az = radec2altaz(comp.position, frame)
                 push!(alts, alt)
@@ -67,10 +67,10 @@ function predict(
                     # MWA Beam is calculated on coarse channels ~ 1.28 MHz
                     # There's no need to calculate the beam values more than once
                     # in this interval.
-                    current_coarse_freq = closest_freq(beam, freq)
+                    current_coarse_freq = closest_freq(beam, Float64(freq))
                     if current_coarse_freq != prior_coarse_freq
                         jones = reshape(
-                            beamjones(beam, freq, alts, azs),
+                            beamjones(beam, Float64(freq), alts, azs),
                             4, length(comps), length(unique_times)
                         )
                         prior_coarse_freq = current_coarse_freq
